lint
SCCS-vsn: sys/kern/subr_rmap.c 4.8
SCCS-vsn: sys/kern/subr_rmap.c.sav 4.8
SCCS-vsn: sys/kern/sys_generic.c 5.21
SCCS-vsn: sys/kern/uipc_mbuf.c 1.40
SCCS-vsn: sys/kern/uipc_socket.c 4.60
SCCS-vsn: sys/vm/vm_swap.c 4.11
SCCS-vsn: sys/vax/vax/conf.c 4.60
SCCS-vsn: sys/vax/vax/machdep.c 4.67

+++ usr/src/sys/kern/subr_rmap.c
@@ -1,4 +1,4 @@
-/*	subr_rmap.c	4.7	82/10/21	*/
+/*	subr_rmap.c	4.8	82/10/21	*/
 
 #include "../h/param.h"
 #include "../h/systm.h"
@@ -43,7 +43,7 @@
  */
 rminit(mp, size, addr, name, mapsize)
 	register struct map *mp;
-	int size, addr;
+	long size, addr;
 	char *name;
 	int mapsize;
 {
